misc(client): moved methods to separate files
