+++ client/src/client.cpp
@@ -8,7 +8,7 @@
 
 
 /**
- * @brief Construct a new rtype::Client::Client object
+ * @brief Constructs a new rtype::Client::Client object
  */
 rtype::Client::Client()
 : _isRunning(true), _start(std::chrono::system_clock::now()), _ecs(initECS()), _graphical(std::make_unique<SFML>()), _currentScene(MAIN_MENU), fps(60), _drawClock(std::chrono::system_clock::now())
@@ -25,7 +25,7 @@ rtype::Client::Client()
 }
 
 /**
- * @brief Destroy the rtype::Client::Client object
+ * @brief Destroys the rtype::Client::Client object
  */
 rtype::Client::~Client()
 {
@@ -33,59 +33,7 @@ rtype::Client::~Client()
 }
 
 /**
- * @brief Initialize the ECS
- *
- * @return ECS
- */
-ECS rtype::Client::initECS()
-{
-    ECS ecs;
-    ecs.registerComponent<Position>();
-    ecs.registerComponent<Health>();
-    ecs.registerComponent<Velocity>();
-    return ecs;
-}
-
-/**
- * @brief Initialize the player and add it to the ECS
- *
- */
-void rtype::Client::initPlayer()
-{
-    _players.push_back(_ecs.createEntity());
-    _ecs.addComponent<Position>(_players[0], {100, 100});
-    _ecs.addComponent<Health>(_players[0], 100);
-    _ecs.addComponent<Velocity>(_players[0], {1, 1, 2});
-}
-
-
-/**
- * @brief maps the keys returned by the graphical lib to the game keys
- *
- * @param keys KeyState returned by the graphical lib
- */
-void rtype::Client::handleKeys(KeyState keys)
-{
-    if(keys.up)
-        std::cout << "UP" << std::endl;
-    if(keys.down)
-        std::cout << "DOWN" << std::endl;
-    if(keys.left)
-        std::cout << "LEFT" << std::endl;
-    if(keys.right)
-        std::cout << "RIGHT" << std::endl;
-    if(keys.space)
-        std::cout << "SPACE" << std::endl;
-    if(keys.shift)
-        std::cout << "SHIFT" << std::endl;
-    if(keys.a)
-        std::cout << "A" << std::endl;
-    if(keys.eight)
-        std::cout << "8" << std::endl;
-}
-
-/**
- * @brief Run the game loop
+ * @brief Runs the game loop
  *
  * @param sender Network::ISender to send data to the server
  * @param receive Network::Receive to receive data from the server
@@ -104,9 +52,6 @@ void rtype::Client::gameLoop(ISender& sender, IReceiver& receive, int port)
 
         if (elapsed > 1000) {
             _start = now;
-            // std::string data = std::to_string(_players[0].getComponent<Position>().x) + " "
-            //                  + std::to_string(_players[0].getComponent<Position>().y) + " "
-            //                  + std::to_string(port);
             std::string data = std::to_string(_ecs.getComponent<Position>(_players[0])->x) + " "
                              + std::to_string(_ecs.getComponent<Position>(_players[0])->y) + " "
                              + std::to_string(port);
